ExtraseEvidentaFinanciara: fix +
I am confused.

It’s my understanding that onClick’s argument here should be a function
of no arguments, or a message data constructor.

Looking at a function of no arguments, it seems to be immediately
evaluated, and so a constant which is computed when the view is
rendered, not when the user clicks the button later at some point.

From how I understand it works, it is not immediately evaluated.

Wait, it begins to make sense: I thought appendView’s first argument the
empty list is captured, and so when I’d later click the button, I’d add
an item to *that* captured empty list, and return the new list with a
single item, although it may have been added a new item from a different
point in the app or some of the list items may have changed. The trick
though is that upon any change to the value, the whole view function is
rerun and it gets the up-to-date list, and onClick would get the correct
thing.

I’m not entirely sure I understand this, but I think I’ll find out later
when I’ll implement item editing.