Quill: receive template ID
I need to pass the document template from Elm to Quill, which I don’t
have in Elm. So I render and fill the template with the initial data in
a <div> with a particular ID and then pass that ID to the JS side of the
port which get’s that DOM’s element inner HTML, and passes it further to
Quill.

Quill then passes back the edited document’s HTML into Elm for storage.