Re-implemented onClick to have a 1-arity handler
One of the issues brought up on elm-discuss user-group [1] was that in
every place where I had an onClick callback I would have my code
pre-compute the after-click state because the onClick callback is defined
as a zero-arity function, which is essentially a constant in Elm.

This may become a problem as my state grows and I get more places where
I need onClick callbacks.

To work around that, I have re-implemented onClick to expect a one-arity
function, and just ignore the one argument I get inside the callback.

[1] https://groups.google.com/forum/#!topic/elm-discuss/ta64y4JBMc4