RezultatIncercareConciliere: fix resetting dropdown
Because the form record changes, the current model is passed into the
dropdown and the selected option is found by comparing that current
model to the option value--which can only work initially when the form
is empty, and so the model can be equal to one of the initial models.

When any of the form fields change though, the new model is no longer
equal to the initial model associated with the dropdown option even if
the type variance matches, and so the dropdown was being reset to the
default option.

The solution is to instead give the dropdown the corresponding *initial*
model for the current model.